Revert "feat: handle partial player fetch failures"
